requirements.txt
# Part 1: Speech-to-Text (Whisper)
openai-whisper>=20231117
scipy>=1.9.0
numpy>=1.20.0
audioread>=3.0.0
yt-dlp>=2024.1.0

# Part 2: Defamation Detector
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.26.0
nltk>=3.8
scikit-learn>=1.0.0
